fix warning
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@2294 a380766d-ff14-0410-b294-a243070f3f08

+++ source/server/SC_World.cpp
@@ -383,7 +383,7 @@ void World_NonRealTimeSynthesis(struct World *world, WorldOptions *inOptions)
 
 		numInputChannels = world->mNumInputs = inputFileInfo.channels; // force it.
 
-		if (inputFileInfo.samplerate != inOptions->mPreferredSampleRate)
+		if (inputFileInfo.samplerate != (int)inOptions->mPreferredSampleRate)
 			scprintf("WARNING: input file sample rate does not equal output sample rate.\n");
 			
 	} else {
